# Optimize GetAssignedExpressionFromDataDeclaration