Test durations:

Test                                     | before feature introduction | master | this fix
-----------------------------------------|-----------------------------|--------|----------
//verilog/formatting:tree_unwrapper_test | 0.2s                        | 6.4s   | 0.2s
//verilog/formatting:formatter_test      | 3.0s                        | 29.2   | 3.1s